Start a Dockerfile
This adds a Dockerfile, to start containerizing the application.

We use the Rocker project's Shiny container, version 4.3.2.  Everything
from the repo is copied into the container, and we use the container's
`myscripts` directory as the working directory.

Then, we install the `remotes` and `readr` packages, which will be
needed by the script that handles package installation.

Finally, as part of the package-install, at the end we clear out `/tmp`,
to make sure temp stuff doesn't end up in the final image.