Start sorting packages into dependencies
With a plain CSV file of packages, it's not enough to just go down the
list, installing the specific versions.  If we do that, eventually we'll
hit a package that cannot install without a dependency that is farther
down the list.  So, start organizing the packages into groupings, with
dependencies places higher up the list.

...

This is definitely not going to work.  I'm new to R, but I know Python,
and R must have an equivalent to lock files.  Let's see if we can find
that.