Dockerfile: Switch to renv
Now that we have an renv lock file, we can use it!

We no longer need to install remotes or readr.  Instead, we bring in
renv.  Then, we restore the environment, which reads our renv.lock file
automatically.

NOTE: This depends on the `COPY` and `WORKDIR` lines that are farher up
in the Dockerfile.