Dockerfile: Set port and run command
This sets the command that will be run by default: Start the Shiny app,
listening on port 3838.

Shiny is set to run on port 3838, and we expose that port to the outside
of the container.  Note that the program running the container is still
responsible for mapping the port to the outside world!