Dockerfile: Set Shiny to listen on all IPs
Shiny listens to localhost by default.  In a running container, chances
are there is a different IP address (not localhost) that's used to
represent communications with outside container.  Localhost inside a
container is used only for communications between things _inside_ the
container (not even for inter-container comms).

So, switch to listening on all IPs.