Dockerfile: Clean up Renv downloads & turn off cache
Renv behavior is to keep copies of downloaded sources, and to store
installed packages in a common location.  This is all done under the
assumption that Renv is being used in a place where multiple
environments will be regularly maintained.  It doesn't hold up in a
situation like this, where each container build is being done in
isolation.

We can turn off installed-package caching, which will have Renv install
the package directory into the R Library path.  But we cannot directly
turn off caching of downloaded files, so we have to clean them up later.