Workflow: Enable SBOM support
Enable recording Software Bill-Of-Materials information in Docker images
we build.  This should include information about the source image, and
information about the Ubuntu packages we install.

NOTE: It probably will not include information about the R packages we
install.

NOTE: It might not include too much source info, depending on if Rocker
builds with SBOM or not.