Add TZ environment variable
This container needs the timezone.  If it doesn't have it, all logs will
use UTC for their log lines.  This can be a problem because the log line
looks like this:

```
INFO [03:12:33] log entry goes here
```

This is the default log format for R's `lgr` logging package.  It shows
the time—in the configured time zone—but not the date.  It also does not
include any indication of the time zone.  That causes a problem with
tools like Splunk.  Splunk, for example, defaults to assuming the log
entry's time zone is the time zone of either the server, or the Splunk
instance.  This can cause Splunk to assume the wrong timestamp.

Setting a default time zone of US/Pacific in the Dockerfile.

The base image already brings in Ubuntu's `tzdata` package; we just need
to set the time zone.